+++ internal/transport/grpc/core/core_test.go
@@ -5,9 +5,9 @@ import (
 	"net"
 	"testing"
 
+	commonv1 "github.com/JekYUlll/Dipole/api/gen/go/common/v1"
+	corev1 "github.com/JekYUlll/Dipole/api/gen/go/core/v1"
 	"github.com/JekYUlll/Dipole/internal/model"
-	commonv1 "github.com/JekYUlll/Dipole/internal/transport/grpc/gen/common/v1"
-	corev1 "github.com/JekYUlll/Dipole/internal/transport/grpc/gen/core/v1"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/credentials/insecure"