+++ internal/transport/grpc/delivery/server_test.go
@@ -7,9 +7,9 @@ import (
 	"testing"
 	"time"
 
+	deliveryv1 "github.com/JekYUlll/Dipole/api/gen/go/delivery/v1"
 	deliverycontract "github.com/JekYUlll/Dipole/internal/realtime/delivery"
 	grpcauth "github.com/JekYUlll/Dipole/internal/transport/grpc/auth"
-	deliveryv1 "github.com/JekYUlll/Dipole/internal/transport/grpc/gen/delivery/v1"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/credentials/insecure"